Refactor #1 - change RouterWithCV.js file path
